Distinguish between developer, reviewer and product owner roles
This commit changes the wording of the yarns so that developers
are distinguished from reviewers and product owners. The tests do not
yet make any distinction internally, as different role-based
access-controls are not configured (there is just 'authorized user' vs
'unauthorized user' at the moment), but this shows the functionality
required in the future.