Adjust config so that per-repo docs can rendered in the future
This change creates a public_html folder per-project, so that formatted
docs can be placed there, and then rendered from that location. It
adjusts the /data directory in the containers to match this structure,
so that the repos are located in /data/gitserver , and the docs can
later be placed in /data/public_html. Finally, it adjusts the cgitrc
scan-path, so that cgit will find the repos at the new location.