Create sphinx directory layout when start.sh runs
This adds a line to the start.sh script so that sphinx 'build' and
'source' directories will be created when the containers first run.
This currently does not check for existing sphinx directories, as
sphinx does this for us; we may wish to add our own check to make
this more robust.

It also copies the sphinx config to an 'immutable' config directory
so we can keep it separate from the source until build time,
in order to regenerate docs with each push to master.