Make the AWS builder work
In AWS the user used to connect will be "fedora" instead of
"root". We can copy files into "/home" with that user, so we
copy the files into "$HOME" which will be /home/fedora, or /root
(aws and qemu/vbox builds respectively).

We also need to set file ownerships for those files, given that
some services (like sssd) require their files to be only readable
by root:root.