Change 'ASSUMING' to 'GIVEN'
In our tests, we use 'ASSUMING' for some initial setup steps. However,
if the implementation of an 'ASSUMING' step should fail for some
reason, yarn will skip the scenario, as per:

http://git.liw.fi/cgi-bin/cgit/cgit.cgi/cmdtest/tree/README.yarn#n123

We should therefore use 'GIVEN' for these steps, as, if a 'GIVEN' step
fails, the scenario, too, will fail.