Remove sample-project-creation line from README
We had a line instructing the user to create a sample project group.
This is no longer necessary; the sample project group is included
on the system, so trying to create it results in an error, as the
group already exists. This patch removes the instruction to create
a sample project group.