Add rudimentary support for user roles to our update hook.
This commit makes our update hook check that a user is in the correct
user groups to perform an action. It is currently very basic, checking
only if a user is allowed to push to (or tag) a branch at all.

This commit also permits the creation of 'hotfix' branches, so that
our infra will allow a user to create a new branch if it has the
'hotfix' prefix.