Introduce workaround for reboot during packer build
Packer doesn't seem to support reboots during the build. Making the
reboot happen after a minute we let some time for packer to finish with
the script and clean it up before the VM goes off. Then the next
provisioning step needs to wait a bit more than the timer to ensure
that at least the VM is off. This step will wait until SSH is available
to continue.