Fix broken haproxy config
Overzealousness when changing the single-letter variable 'i' to
'project' meant that a literal 'i' was caught in the excitement.
As a result, our haproxy config was invalid, having the
meaningless '-project' option for the ACL, rather than the more
conventional '-i'. This patch changes '-project' back to '-i'
in the relevant place.